Address comments
Create utils file for import script
Break script in smaller helper functions
Use lodash more extensively
